langgraph_task_agent/requirements.txt
langgraph
boto3
langchain_community
langchain-aws
python-dotenv
